Fix import cycle
