+++ mypyc/genops.py
@@ -34,7 +34,6 @@ from mypy.types import (
     Type, Instance, TupleType, AnyType, TypeOfAny, UninhabitedType, get_proper_type
 )
 from mypy.visitor import ExpressionVisitor, StatementVisitor
-from mypy.state import strict_optional_set
 from mypy.util import split_target
 
 from mypyc.common import TEMP_ATTR_NAME, TOP_LEVEL_NAME
@@ -43,7 +42,7 @@ from mypyc.ops import (
     BasicBlock, AssignmentTarget, AssignmentTargetRegister, AssignmentTargetIndex,
     AssignmentTargetAttr, AssignmentTargetTuple, Environment, LoadInt, RType, Value, Register, Op,
     FuncIR, Assign, Branch, RTuple, Unreachable,
-    TupleGet, ClassIR, NonExtClassInfo, RInstance, ModuleIR, ModuleIRs, GetAttr, SetAttr,
+    TupleGet, ClassIR, NonExtClassInfo, RInstance, ModuleIR, GetAttr, SetAttr,
     LoadStatic, InitStatic, INVALID_FUNC_DEF, int_rprimitive,
     bool_rprimitive, list_rprimitive, is_list_rprimitive, dict_rprimitive, set_rprimitive,
     str_rprimitive, none_rprimitive, is_none_rprimitive, object_rprimitive,
@@ -65,8 +64,6 @@ from mypyc.nonlocalcontrol import (
 )
 from mypyc.genopscontext import FuncInfo, ImplicitClass
 from mypyc.genopsmapper import Mapper
-from mypyc.genopsvtable import compute_vtable
-from mypyc.genopsprepare import build_type_map
 from mypyc.ir_builder import LowLevelIRBuilder
 from mypyc.specialize import specialize_function
 
@@ -79,62 +76,6 @@ class IRVisitor(ExpressionVisitor[Value], StatementVisitor[None]):
 
 class UnsupportedException(Exception):
     pass
-
-
-# The stubs for callable contextmanagers are busted so cast it to the
-# right type...
-F = TypeVar('F', bound=Callable[..., Any])
-strict_optional_dec = cast(Callable[[F], F], strict_optional_set(True))
-
-
-@strict_optional_dec  # Turn on strict optional for any type manipulations we do
-def build_ir(modules: List[MypyFile],
-             graph: Graph,
-             types: Dict[Expression, Type],
-             mapper: 'Mapper',
-             options: CompilerOptions,
-             errors: Errors) -> ModuleIRs:
-
-    build_type_map(mapper, modules, graph, types, options, errors)
-
-    result = OrderedDict()  # type: ModuleIRs
-
-    # Generate IR for all modules.
-    class_irs = []
-
-    for module in modules:
-        # First pass to determine free symbols.
-        pbv = PreBuildVisitor()
-        module.accept(pbv)
-
-        # TODO: Get rid of cyclic dependency by extractring build_ir to a new module
-        from mypyc.genopsvisitor import IRBuilderVisitor
-        visitor = IRBuilderVisitor()
-
-        # Second pass.
-        builder = IRBuilder(
-            module.fullname, types, graph, errors, mapper, pbv, visitor, options
-        )
-
-        visitor.builder = builder
-
-        builder.visit_mypy_file(module)
-        module_ir = ModuleIR(
-            module.fullname,
-            list(builder.imports),
-            builder.functions,
-            builder.classes,
-            builder.final_names
-        )
-        result[module.fullname] = module_ir
-        class_irs.extend(builder.classes)
-
-    # Compute vtables.
-    for cir in class_irs:
-        if cir.is_ext_class:
-            compute_vtable(cir)
-
-    return result
 
 
 class IRBuilder:
